Split Dockerfile into layers
Keeping both layers in the same file. This way they
can be referenced within the file without knowing the
final tag they will be pushed under.
The internal names are "base" and "test".

Build using (for example):
docker build --target build --tag Fedora-36-base
docker build --target test --tag Fedora-36-test

It is currently not possible to build and tag both layers
in one build command. But that should not be a problem,
the "build" (base) image is cached by docker and reused in the
next call to `docker build`.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>