Add images based on Fedora 36 (#1)
* Add image based on Fedora 36

Add a Dockerfile for an image based on Fedora 36, containing gcc 12,
qemu, and other tools needed to run the EDK2 CI.

Also put GitHub Action in place that builds the image and pushes it to
the GitHub image registry (ghcr.io). The image shows up under "Packages"
on the GitHub page as well. The image is only pushed for changes on the
main branch. Pull Requests build the image, but do not push it.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>

* Split Dockerfile into layers

Keeping both layers in the same file. This way they
can be referenced within the file without knowing the
final tag they will be pushed under.
The internal names are "base" and "test".

Build using (for example):
docker build --target build --tag Fedora-36-base
docker build --target test --tag Fedora-36-test

It is currently not possible to build and tag both layers
in one build command. But that should not be a problem,
the "build" (base) image is cached by docker and reused in the
next call to `docker build`.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>

* CI: Build and push both layers/images

Calling docker build and docker push directly via bash script.
Not sure how to do that with a GitHub action.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>

* Add comments and license statement to Dockerfile

Signed-off-by: Oliver Steffen <osteffen@redhat.com>

* Add license statement to workflow file

Signed-off-by: Oliver Steffen <osteffen@redhat.com>