Introduce ubuntu20 build and dev images
These are similar to the existing fedora images, but based on ubuntu20.

The build image is similar in spirit to the fedora build image.  It can
be used in automation.

A test image hasn't been introduced yet.

A dev image is included and is intended for direct use by an EDK2
developer.  To use it, the developer must mount their home directory
into the container.  They may also mount the EDK2 workspace or simply
place it in their home directory.