Add folder structure to allow multiple images (#16)
Add a folder structure to allow multiple Dockerfiles/Images
in the repo, and add a common GitHub Workflow to build images.

Each Dockerfile lives in its own folder. The folders
are the same as the image name. Each image has its own
GitHub Workflow yaml file in ".github/workflow/".
The workflow file calls the common image-building
workflow in build-images.yaml, and passes the
image name and a list of sub-image names (aka targets,
for example "build" or "test") to it.

The workflow tags the images like this:
ghcr.io/tianocore/containers/<image name>-<target>:<commit sha>