Add EDK2 test build
Run a build of OvmfPkg/ArmVirtPkg on images created via the GitHub workflow.
This serves as a very basic smoke-test to see if the imge is able to
build EDK2.

It uses stuart to build OvmfPkg for X64, and ArmVirtPkg for AARCH64 and
ARM.

For now, only the last sub-image (aka target) in the list for each
image is tested.

Note: contains a workaround. The external dependency files or AARCH64 and
ARM compilers are removed, so that the ones supplied by the image are
used. Currently, stuart on the EDK2 master branch downloads all tools
at runtime. This workaround can be removed later.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>