Add sudo to fedora to allow tool installation in azure pipelines (#28)
Add sudo tool to baseline of fedora image. This is required to allow non-root users, such as the one created by azure pipelines, to run command requiring root, such as microdnf or rpm to install customized pipeline tools. This is needed as some pipelines leverage powershell or other tools that aren't needed as part of the core build requirements.

Issue #27