Fedora-35: Add Powershell to build image (#52)
The build image is used for building in pipelines. Pipelines are
executed in Azure DevOps and most commonly used Azure-hosted VM
images. Those images come with Powershell installed.

Example:
https://github.com/actions/runner-images/blob/ubuntu22/20230109.1/images/linux/Ubuntu2204-Readme.md#powershell-tools

Further, the Azure Pipelines PowerShell@2 task supports Linux, macOS,
and Windows.

https://learn.microsoft.com/en-us/azure/devops/pipelines/tasks/reference/powershell-v2?view=azure-pipelines

Therefore, pipeline scripts written in Powershell are already built
for cross-platform support.

This change adds Powershell to the build container image so it is
available within build environments.

Based on local testing, this increased the container image size from:
  1.55GB to 1.79GB

This seemed reasonable for the increase in functionality that is
enabled.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>