As with Ubuntu containers, set up a user to match outside user
As with the Ubuntu containers, add an entrypoint script that sets up a
user which matches that which is being used to run the container.

However, if EDK2_DOCKER_USER_HOME hasn't been set, just run the
container as root instead of failing.

Add the user to the 'wheel' group and set the password to 'tianocore' to
allow them to run commands as root via sudo.

Signed-off-by: Rebecca Cran <rebecca@bsdio.com>