Upgrade to Qemu 8, Bump GCC minor version on Fedora  (#72)
* Fedora-35: upgrade to Qemu 8.0.0

Upgrade Qemu to version 8.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>

* Fedora-37: upgrade to Qemu 8.0.0

Upgrade Qemu to version 8.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>

* Fedora 37: Bump gcc to 12.3.1

Fedora 37 moved to gcc 12.3.1. Since we specify exact versions in the
Dockerfile we need to update it there too.

Versions for the corss-compiler stay the same.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>

* Fedora 35: Bump gcc to 11.3.1

Fedora 35 moved to gcc 11.3.1. Since we specify exact versions in the
Dockerfile we need to update it there too.

Version number for the cross-compiler stay the same.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>

---------

Signed-off-by: Oliver Steffen <osteffen@redhat.com>