Ubuntu-22: Update entrypoint to run as non-root on macOS
On macOS (and Windows), docker containers run as a more heavyweight VM
unlike on Linux. One effect of this is that the code in the entrypoint
script will see the UID and GID of the home directory as 0 and try to
create a user with UID 0.

Work around this limitation by allowing users to pass in the UID and
GID when running the container. Update the Readme.md to suggest
using `$(id -u)` and `$(id -g)`.

Signed-off-by: Rebecca Cran <rebecca@bsdio.com>