Remove Ubuntu-20 image
Remove the Ubuntu 20.04 image, since the distribution is very
old and this image is not maintainable anymore.
Note: The existing image will still be available on ghcr.

The availble python version (3.9) is too old for current EDK2.
The edk2-pytools are not available for it:

$ python --version
Python 3.9.5

python -m pip install --upgrade -r pip-requirements.txt
ERROR: Ignored the following versions that require a different python version:
      0.19.0 Requires-Python >=3.10;
      0.19.1 Requires-Python >=3.10;
      0.19.2 Requires-Python >=3.10;
      0.19.3 Requires-Python >=3.10;
      0.19.4 Requires-Python >=3.10;
      0.19.5 Requires-Python >=3.10;
      0.19.6 Requires-Python >=3.10;
      0.19.7 Requires-Python >=3.10;
      0.19.8 Requires-Python >=3.10;
      0.19.9 Requires-Python >=3.10;
      0.20.0 Requires-Python >=3.10;
      0.21.2 Requires-Python >=3.10;
      0.21.3 Requires-Python >=3.10;
      0.21.4 Requires-Python >=3.10;
      0.21.5 Requires-Python >=3.10
ERROR: Could not find a version that satisfies the requirement edk2-pytool-library==0.20.04
  (from versions: 0.9.0, 0.9.1, 0.9.2, 0.10.0, 0.10.1, 0.10.2, 0.10.3, 0.10.4, 0.10.5,
  0.10.6, 0.10.7, 0.10.8, 0.10.9, 0.10.10, 0.10.11, 0.10.12, 0.10.13, 0.10.15, 0.11.0,
  0.11.1, 0.11.2, 0.11.4, 0.11.5, 0.11.6, 0.12.0, 0.12.1, 0.12.2, 0.13.0, 0.13.1, 0.14.0,
  0.14.1, 0.15.0, 0.15.1, 0.15.2, 0.15.3, 0.15.4, 0.16.0, 0.16.1, 0.16.2, 0.17.0, 0.18.0,
  0.18.1, 0.18.2)
ERROR: No matching distribution found for edk2-pytool-library==0.20.0

Signed-off-by: Oliver Steffen <osteffen@redhat.com>