Add Fedora 43 AArch64 image and fix OvmfPkg in AArch64 images (#121)
* Minor cleanups to Fedora-41-AArch64.yaml

'**' branch filter is presumably left-over debug code. Changed to main to
match other container .yaml files.

Removed additional whitespace to match all other .yaml files,
specifically, to make diffs less noisy.

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>

* Get OvmfPkg working in Fedora-41-AArch64

This is simply a matter of adding some non-standard (compared to
other EDK2 packages) gcc prefix environment variables.

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>

* Add Fedora-43-AArch64 image

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>

* Deduplicate AArch64 image logic

Dockerfile and build.yaml are now shared.

With regard to the Dockerfile changes:

gcc-x86_64-linux-gnu was not included before in amd64 (x86_64) images.
It is not needed as long as the additional prefix variables for gcc
(which are required for the arm64 image) are not set. If they are set,
then gcc will be searched for with its full prefixed name and not
found. Unfortunately there is no clean logic for conditionally running
ENV commands in Dockerfiles, so this was considered the cleanest choice.

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>

---------

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>